feat: add tos page
