+++ apps/io-sign-backoffice-app/src/app/[locale]/terms-of-service/page.tsx
@@ -1,0 +1,10 @@
+import PolicyPage from "@/components/PolicyPage";
+
+export default function TOS() {
+  return (
+    <PolicyPage
+      title="Termini e condizioni d'uso"
+      policyId="c6eba0ac-6e1a-4972-b625-930696f7a074"
+    />
+  );
+}
